+++ admin/students/index.php
@@ -46,7 +46,7 @@
                 <td><?= $row["email"]; ?></td>
                 <td><?= $row["phone"]; ?></td> 
                 <td><?= $row["faculty"]; ?></td> 
-                <td><a href="ubahSis.php?id=<?= $row["studentId"]; ?>">ubah</a> || <a href="hapusSis.php?id=<?= $row["studentId"]; ?>"> Hapus</a></td>
+                <td><a href="ubahSis.php?id=<?= $row["studentId"]; ?>">ubah</a> || <a href="hapusSis.php?id=<?= $row["studentId"]; ?>" onclick="return confirm('Apakah kamu yakin ingin menghapus data ini?')"> Hapus</a></td>
             </tr>
         <?php $i++; ?>
         <?php endforeach; ?>
